# Swaggerapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for a course management system.",
        "title": "Go API with Swagger",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/courses": {
            "get": {
                "description": "Retrieves a list of all courses",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all courses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Course"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new course to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new course",
                "parameters": [
                    {
                        "description": "Course data to create",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Course"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created course",
                        "schema": {
                            "$ref": "#/definitions/main.Course"
                        }
                    },
                    "400": {
                        "description": "No data inside the request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/courses/{course_id}": {
            "get": {
                "description": "Get details for a specific course by its unique ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a single course by ID",
                "operationId": "get-one-course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Course"
                        }
                    },
                    "404": {
                        "description": "No course found with the given ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the details of a course by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update an existing course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated course data",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Course"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Course"
                        }
                    },
                    "404": {
                        "description": "No course found with the given ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a course by its unique ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "No course found with the given ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Author": {
            "type": "object",
            "properties": {
                "fullname": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "main.Course": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/main.Author"
                },
                "course_id": {
                    "type": "string"
                },
                "course_name": {
                    "type": "string"
                },
                "course_price": {
                    "type": "integer"
                }
            }
        }
    }
}